+++ src/core/core.svelte
@@ -166,6 +166,23 @@
     }
   };
   const onContentTouchStart = (e) => {
+    // skip inputs
+    let isInputElement = e.target.tagName === 'INPUT' || e.target.tagName === 'TEXTAREA';
+    if (isInputElement) {
+      return;
+    }
+    // skip scrollable elements
+    let isScrollElement = false;
+    if (typeof window.getComputedStyle === 'function') {
+      const style = window.getComputedStyle(e.target);
+      if (style.overflow === 'auto' || style.overflow === 'scroll') {
+        isScrollElement = true;
+      }
+    }
+    if (isScrollElement) {
+      // (window as any)._vcOrigConsole.log('onContentTouchStart isScrollElement', isScrollElement);
+      return;
+    }
     const top = divContent.scrollTop,
           totalScroll = divContent.scrollHeight,
           currentScroll = top + divContent.offsetHeight;
@@ -177,26 +194,23 @@
       // scrollTop always equals to 0 (it is always on the top),
       // so we need to prevent scroll event manually
       if (divContent.scrollTop === 0) {
-        if (e.target.classList && !e.target.classList.contains('vc-cmd-input')) { // skip input
-          preventContentMove = true;
-        }
+        preventContentMove = true;
       }
     } else if (currentScroll === totalScroll) {
       // when content is on the bottom,
       // do similar processing
       divContent.scrollTop = top - 1;
       if (divContent.scrollTop === top) {
-        if (e.target.classList && !e.target.classList.contains('vc-cmd-input')) {
-          preventContentMove = true;
-        }
+        preventContentMove = true;
       }
     }
-    // (window as any)._vcOrigConsole.log('preventContentMove', preventContentMove);
+    // (window as any)._vcOrigConsole.log('onContentTouchStart preventContentMove', preventContentMove);
   };
   const onContentTouchMove = (e) => {
     if (preventContentMove) {
       e.preventDefault();
     }
+    // (window as any)._vcOrigConsole.log('onContentTouchMove preventContentMove', preventContentMove);
   };
   const onContentTouchEnd = (e) => {
     preventContentMove = false;
@@ -236,13 +250,21 @@
     },
     touchMove(e) {
       const touch = e.changedTouches[0];
-      if (Math.abs(touch.pageX - mockTapInfo.touchstartX) > mockTapInfo.tapBoundary || Math.abs(touch.pageY - mockTapInfo.touchstartY) > mockTapInfo.tapBoundary) {
+      if (
+        Math.abs(touch.pageX - mockTapInfo.touchstartX) > mockTapInfo.tapBoundary || 
+        Math.abs(touch.pageY - mockTapInfo.touchstartY) > mockTapInfo.tapBoundary
+      ) {
         mockTapInfo.touchHasMoved = true;
       }
+      // (window as any)._vcOrigConsole.log('mockTapEvent.touchMove',  mockTapInfo.touchHasMoved);
     },
     touchEnd(e) {
       // move and time within limits, manually trigger `click` event
-      if (mockTapInfo.touchHasMoved === false && e.timeStamp - mockTapInfo.lastTouchStartTime < mockTapInfo.tapTime && mockTapInfo.targetElem != null) {
+      if (
+        mockTapInfo.touchHasMoved === false && 
+        e.timeStamp - mockTapInfo.lastTouchStartTime < mockTapInfo.tapTime && 
+        mockTapInfo.targetElem != null
+      ) {
         const tagName = mockTapInfo.targetElem.tagName.toLowerCase();
         let needFocus = false;
         switch (tagName) {