+++ CHANGELOG.md
@@ -7,6 +7,7 @@ English | [简体中文](./CHANGELOG_CN.md)
 - `Feat(Network)` Add recycle scrolling to imporove performance.
 - `Feat(Network)` Add "Start Time" of a request.
 - `Feat(Network)` Use `curl` instead of `url` as the copy value of a request. (issue #410)
+- `Fix(Storage)` Fix an event bug that overflow content cannot scroll. (issue #542)
 
 
 ## 3.14.7 (2022-09-23)